+++ CHANGELOG.md
@@ -9,8 +9,10 @@ This document is based on Keep a Changelog 1.0.0
 ### Changed
 - Changed how saves and objects are handled internally.
 - Improved particle flows performances when using a big amount of particles.
+- Optimized bar visualizers when changing their parameters.
 - Changing an object's name is now done using a dedicated field in the rack (less clicks!).
 - Changed rules for timers positioning. (The converter should reajust timers when opening a project from an older versions.)
+- Changed some default values (it only affects newly created objects and new object parameters).
 - Irrelevant options (i.e that have no effect) in some objects are no longer available (timers, visualizers). In consequence they are no longer stored as well with these objects (as they are useless).
 
 ## 0.2.2-beta (2021-07-28)